// front-end/dashboard/src/api-client/models/Image.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Task Management Service
 * API for managing tasks for patients, doctors, and administrators.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Image
 */
export interface Image {
    /**
     * 
     * @type {string}
     * @memberof Image
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof Image
     */
    type?: string;
    /**
     * 
     * @type {string}
     * @memberof Image
     */
    signedUrl?: string;
    /**
     * 
     * @type {string}
     * @memberof Image
     */
    url?: string;
    /**
     * 
     * @type {string}
     * @memberof Image
     */
    patientId?: string;
    /**
     * 
     * @type {string}
     * @memberof Image
     */
    taskId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Image
     */
    createdAt?: string;
    /**
     * 
     * @type {string}
     * @memberof Image
     */
    updatedAt?: string;
}

/**
 * Check if a given object implements the Image interface.
 */
export function instanceOfImage(value: object): value is Image {
    return true;
}

export function ImageFromJSON(json: any): Image {
    return ImageFromJSONTyped(json, false);
}

export function ImageFromJSONTyped(json: any, ignoreDiscriminator: boolean): Image {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'type': json['type'] == null ? undefined : json['type'],
        'signedUrl': json['signedUrl'] == null ? undefined : json['signedUrl'],
        'url': json['url'] == null ? undefined : json['url'],
        'patientId': json['patientId'] == null ? undefined : json['patientId'],
        'taskId': json['taskId'] == null ? undefined : json['taskId'],
        'createdAt': json['createdAt'] == null ? undefined : json['createdAt'],
        'updatedAt': json['updatedAt'] == null ? undefined : json['updatedAt'],
    };
}

export function ImageToJSON(json: any): Image {
    return ImageToJSONTyped(json, false);
}

export function ImageToJSONTyped(value?: Image | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'type': value['type'],
        'signedUrl': value['signedUrl'],
        'url': value['url'],
        'patientId': value['patientId'],
        'taskId': value['taskId'],
        'createdAt': value['createdAt'],
        'updatedAt': value['updatedAt'],
    };
}